requirements.txt
# packaging
pip>=9
setuptools>=26
twine
wheel>=0.29

# data science
numpy
pandas
scikit-learn

# tests/linter
black
coverage
flake8
pytest
yapf

# API
gcsfs
google-cloud-storage
mlflow
s3fs
fastapi
uvicorn

# utilities
six>=1.14
joblib
memoized-property
termcolor
nltk
<<<<<<< HEAD
=======

#twitter_api
python-dotenv
tweepy

# deep learning
tensorflow==2.8.0
gensim
>>>>>>> 55a970c46a555ee90cda9250d0ff3bc529ba068e
